Fix comments and typos.
