+++ src/pydvl/value/shapley/gt.py
@@ -141,7 +141,7 @@ def _group_testing_shapley(
         progress: Whether to display progress bars for each job.
         job_id: id to use for reporting progress (e.g. to place progres bars)
         seed: Either an instance of a numpy random number generator or a seed
-        for it.
+            for it.
     Returns:
 
     """
@@ -203,7 +203,7 @@ def group_testing_shapley(
             address, number of cpus, etc.
         progress: Whether to display progress bars for each job.
         seed: Either an instance of a numpy random number generator or a seed
-        for it.
+            for it.
         options: Additional options to pass to
             [cvxpy.Problem.solve()](https://www.cvxpy.org/tutorial/advanced/index.html#solve-method-options).
             E.g. to change the solver (which defaults to `cvxpy.SCS`) pass
